+++ apps/block_scout_web/lib/block_scout_web/controllers/address_token_balance_controller.ex
@@ -2,7 +2,7 @@ defmodule BlockScoutWeb.AddressTokenBalanceController do
   use BlockScoutWeb, :controller
 
   alias BlockScoutWeb.AccessHelpers
-  alias Explorer.{Chain, Market}
+  alias Explorer.Chain
   alias Explorer.Chain.Address
   alias Indexer.Fetcher.TokenBalanceOnDemand
 
@@ -17,10 +17,6 @@ defmodule BlockScoutWeb.AddressTokenBalanceController do
         TokenBalanceOnDemand.trigger_fetch(address_hash, token_balances)
       end)
 
-      token_balances_with_price =
-        token_balances
-        |> Market.add_price()
-
       case AccessHelpers.restricted_access?(address_hash_string, params) do
         {:ok, false} ->
           conn
@@ -28,7 +24,7 @@ defmodule BlockScoutWeb.AddressTokenBalanceController do
           |> put_layout(false)
           |> render("_token_balances.html",
             address_hash: Address.checksum(address_hash),
-            token_balances: token_balances_with_price,
+            token_balances: token_balances,
             conn: conn
           )